cleanup shop functions
